All commands responses can be a failure response
A failure response is either unsupported or (error "msg") in
smtlib. Previously, only the GenResponse could return one of
those. This was a mistake, as a solver might respond unsupported
or error to any command. Unsupported if for example an option set with
set-option is not supported, or error if some operation in the asserted
formulas are not supported.

This refactoring keep the previous response types, but make all of them
as abstract traits that are inherited by a concrete successful response
for each kind, as well as Unsupported and Error.